Move plan modifiers to their own files
